+++ src/Aks/Aks.Test/ScenarioTests/AksTestRunner.cs
@@ -66,7 +66,7 @@ namespace Commands.Aks.Test.ScenarioTests
                     {
                         AzureSession.Instance.DataStore = new MemoryDataStore();
                         var home = Environment.GetFolderPath(Environment.SpecialFolder.UserProfile);
-                        var dir = Path.GetDirectoryName(new Uri(Assembly.GetExecutingAssembly().CodeBase).AbsolutePath);
+                        var dir = Path.GetDirectoryName(new Uri(Assembly.GetExecutingAssembly().Location).AbsolutePath);
                         var subscription = HttpMockServer.Variables["SubscriptionId"];
                         AzureSession.Instance.DataStore.WriteFile(Path.Combine(home, ".ssh", "id_rsa.pub"), File.ReadAllText(dir + "/Fixtures/id_rsa.pub"));
                         var jsonOutput = @"{""" + subscription + @""":{ ""service_principal"":""foo"",""client_secret"":""bar""}}";
@@ -76,7 +76,7 @@ namespace Commands.Aks.Test.ScenarioTests
                     {
                         AzureSession.Instance.DataStore = new MemoryDataStore();
                         var home = Environment.GetFolderPath(Environment.SpecialFolder.UserProfile);
-                        var dir = Path.GetDirectoryName(new Uri(Assembly.GetExecutingAssembly().CodeBase).AbsolutePath);
+                        var dir = Path.GetDirectoryName(new Uri(Assembly.GetExecutingAssembly().Location).AbsolutePath);
                         
                         var currentEnvironment = TestEnvironmentFactory.GetTestEnvironment();
                         var subscription = currentEnvironment.ConnectionString.KeyValuePairs["SubscriptionId"];