+++ src/Storage/Storage/Common/AzureStorageFileDirectory.cs
@@ -145,6 +145,10 @@ namespace Microsoft.WindowsAzure.Commands.Common.Storage.ResourceModel
         // Convert Track2 File Dir object to Track 1 file Dir object
         public static CloudFileDirectory GetTrack1FileDirClient(ShareDirectoryClient shareFileDirClient, StorageCredentials credentials)
         {
+            if (credentials.IsToken)
+            {
+                return new InvalidCloudFileDirectory(shareFileDirClient.Uri, credentials);
+            }
             if (credentials.IsSAS) // the Uri already contains credentail.
             {
                 credentials = null;