+++ src/Databricks/ChangeLog.md
@@ -19,6 +19,11 @@
 -->
 ## Upcoming Release
 
+## Version 1.7.0
+* Added some parameters in `Update-AzDatabricksWorkspace`:
+    - `EnableNoPublicIP`
+    - `PublicNetworkAccess`
+
 ## Version 1.6.0
 * Added some parameters in the `New-AzDatabricksWorkspace` and `Update-AzDatabricksWorkspace`.
     - `ManagedDiskKeyVaultPropertiesKeyName`