+++ src/PolicyInsights/PolicyInsights/ChangeLog.md
@@ -18,6 +18,10 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* Updated Azure.Core to 1.34.0.
+
+## Version 1.6.2
+* Updated Azure.Core to 1.33.0.
 
 ## Version 1.6.1
 * Updated Azure.Core to 1.31.0.